chore: renamed stellar API file
